M20 phase 3: x86_64 HAL surface (GDT, IDT, TSS, ISR stubs, switch.s)
Brings up the per-arch HAL files under kernel/hal/x86_64/ that
mirror everything in kernel/hal/x86/.  All 9 files compile cleanly
under `make ARCH=x86_64`; full-kernel linking is deferred to Phase 5
once the VMM is ported.

Files added:
- io.c          — port I/O + hal_reboot/hal_shutdown.  Word-for-word
                   port of i386; in/out opcodes are identical in
                   long mode.
- hal_arch.c    — hal_api.h impl.  Same shape as i386 with rflags/
                   pushfq/popq instead of eflags/pushf/pop.
                   hal_syscall_exit_to_kernel is stubbed (Phase 7
                   lands SYSCALL/SYSRET).
- gdt.c         — 7-slot 64-bit GDT (null + kernel CS64/DS + user
                   CS64/DS + 16-byte TSS desc spanning slots 5+6).
                   Code segments use L=1, D=0 per AMD64 APM §4.8.
                   gdtr is 10 bytes (2 limit + 8 base) for lgdt.
- tss.c         — 64-bit TSS.  Packed struct because RSP0 starts at
                   offset 4 (not 8) and the whole layout is unaligned.
                   RSP0 set to a dedicated 4 KiB syscall_stack.
- idt.c         — 64-bit IDT with 16-byte gate descriptors.  PIC
                   remap, IRQ install, APIC routing all carry over
                   unchanged from i386.  isr_handler walks rip/cs/
                   err_code from the int_frame using the new field
                   names.
- isr_stubs.s   — Per-vector stubs + isr_common.  Pushes 15 GPRs
                   (rax..r15 minus rsp), no segment register save
                   (long mode largely ignores ds/es/fs/gs).  Calls
                   isr_handler with frame in rdi per System V ABI.
                   iretq instead of iret.  Stack alignment verified
                   16-byte at call point.
- switch.s      — context_switch(uint64_t* save_rsp_to, uint64_t
                   new_rsp).  Saves rbx, rbp, r12-r15 per System V.
                   Args in rdi/rsi (not stack).
- task_arch.c   — hal_task_init_stack lays down the first-switch
                   frame matching switch.s pop order.  task_trampoline
                   reads entry from rbx (saved as 4th pop).

Header changes (both archs):
- kernel/includes/idt.h — struct int_frame is now arch-conditional
  via #if defined(__x86_64__).  i386 layout unchanged; x86_64 layout
  has 15 GPRs + 5-quadword CPU-pushed frame.  Field names int_no /
  err_code identical across archs so portable IRQ handlers compile
  on both.
- kernel/includes/tss.h — tss_set_kernel_stack / tss_get_addr widened
  to uintptr_t for pointer hygiene.  No external callers so the
  change is source-compatible on i386 (uintptr_t == uint32_t).
- kernel/hal/x86/tss.c — signature updated to match the new
  uintptr_t API.  Behavior unchanged.

Makefile:
- x86_64 ARCH_C_SRCS / ARCH_ASM_SRCS grow to include the 9 new
  files.  Note that linking still fails on x86_64 (kernel core
  references kprintf, schedule_check, task_exit, etc. that are
  not in the source list yet) — Phase 5 closes the gap by adding
  the full CORE_C_SRCS for x86_64 too.

i386 baseline rebuild + boot tested: all M19/M18.5 self-tests still
PASS, shell prompt up.

x86_64 verification: all 9 .o files compile cleanly with -Wall
-Wextra under `make ARCH=x86_64`.  Link target reports the expected
unresolved symbols from kernel_main's deps.

M20-PROGRESS.md
- Phase 3 status flipped to shipped.
- Lessons learned: int_frame arch-conditional pattern, 16-byte
  long-mode TSS descriptor + 2-slot GDT span, uniform 5-quadword
  CPU push on long-mode interrupt entry, missing header deps in
  the Makefile (workaround: clean rebuild).
- Next concrete action updated to Phase 4 (VMM port + vmm.h
  uintptr_t widening).

+++ kernel/hal/x86/tss.c
@@ -63,9 +63,9 @@ void tss_init(void) {
     tss.iomap_base = sizeof(tss);
 }
 
-void tss_set_kernel_stack(uint32_t esp) {
-    tss.esp0 = esp;
+void tss_set_kernel_stack(uintptr_t esp) {
+    tss.esp0 = (uint32_t)esp;
 }
 
-uint32_t tss_get_addr(void)  { return (uint32_t)(uintptr_t)&tss; }
-uint32_t tss_get_limit(void) { return sizeof(tss) - 1; }
+uintptr_t tss_get_addr(void)  { return (uintptr_t)&tss; }
+uint32_t  tss_get_limit(void) { return sizeof(tss) - 1; }
